[HW14] Refactored storage
