+++ hw12_13_14_15_calendar/internal/storage/memory/storage.go
@@ -6,10 +6,11 @@ import (
 
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/configs"
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage/basic"
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage/memory/events"
 )
 
 type Storage struct {
-	events *EventRepository
+	events *events.EventRepository
 	cfg    configs.StorageConf
 	mu     *sync.RWMutex
 }
@@ -17,7 +18,7 @@ type Storage struct {
 func New(cfg configs.StorageConf) *Storage {
 	m := sync.RWMutex{}
 	return &Storage{
-		events: &EventRepository{mu: &m},
+		events: &events.EventRepository{M: &m},
 		cfg:    cfg,
 		mu:     &m,
 	}
