+++ hw12_13_14_15_calendar/internal/storage/sql/storage.go
@@ -9,17 +9,18 @@ import (
 	_ "github.com/lib/pq"
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/configs"
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage/basic"
+	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage/sql/events"
 )
 
 type Storage struct {
 	cfg    configs.StorageConf
 	db     *sqlx.DB
-	events *EventRepository
+	events *events.EventRepository
 }
 
 func New(cfg configs.StorageConf) *Storage {
 	return &Storage{
-		events: &EventRepository{},
+		events: &events.EventRepository{},
 		cfg:    cfg,
 	}
 }
@@ -32,7 +33,7 @@ func (s *Storage) Init(ctx context.Context) error {
 	db, err := sqlx.ConnectContext(ctx, s.cfg.Driver, s.cfg.Conn)
 	if err == nil {
 		s.db = db
-		s.events.db = s.db
+		s.events.DB = s.db
 	}
 	return err
 }
